[IMP] Moved ac_api_requires_client_access into namespace blueprints
