+++ source/app/blueprints/rest/manage/manage_customers_routes.py
@@ -43,7 +43,7 @@ from app.iris_engine.utils.tracker import track_activity
 from app.models.authorization import Permissions
 from app.schema.marshables import ContactSchema
 from app.schema.marshables import CustomerSchema
-from app.util import ac_api_requires_client_access
+from app.blueprints.access_controls import ac_api_requires_client_access
 from app.util import response_error
 from app.util import response_success
